New features, reports and fix bugs

+++ src/main/java/br/com/diocesesjc/mesce/service/PessoaParoquiaService.java
@@ -1,8 +1,10 @@
 package br.com.diocesesjc.mesce.service;
 
 import br.com.diocesesjc.mesce.converter.PessoaParoquiaConverter;
+import br.com.diocesesjc.mesce.converter.data.PessoaConverter;
 import br.com.diocesesjc.mesce.dtos.PessoaParoquiaRequest;
 import br.com.diocesesjc.mesce.dtos.response.PessoaParoquiaResponse;
+import br.com.diocesesjc.mesce.dtos.response.PessoaResponse;
 import br.com.diocesesjc.mesce.entity.Paroquia;
 import br.com.diocesesjc.mesce.entity.Pessoa;
 import br.com.diocesesjc.mesce.entity.PessoaParoquia;
@@ -11,7 +13,11 @@ import br.com.diocesesjc.mesce.repository.ParoquiaRepository;
 import br.com.diocesesjc.mesce.repository.PessoaParoquiaRepository;
 import br.com.diocesesjc.mesce.repository.PessoaRepository;
 import br.com.diocesesjc.mesce.repository.UsuarioRepository;
+import java.math.BigInteger;
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -20,13 +26,15 @@ public class PessoaParoquiaService {
 
     private final PessoaParoquiaRepository pessoaParoquiaRepository;
     private final PessoaParoquiaConverter pessoaParoquiaConverter;
+    private final PessoaConverter pessoaConverter;
     private final PessoaRepository pessoaRepository;
     private final ParoquiaRepository paroquiaRepository;
     private final UsuarioRepository usuarioRepository;
 
-    public PessoaParoquiaService(PessoaParoquiaRepository pessoaParoquiaRepository, PessoaParoquiaConverter pessoaParoquiaConverter, PessoaRepository pessoaRepository, ParoquiaRepository paroquiaRepository, UsuarioRepository usuarioRepository) {
+    public PessoaParoquiaService(PessoaParoquiaRepository pessoaParoquiaRepository, PessoaParoquiaConverter pessoaParoquiaConverter, PessoaConverter pessoaConverter, PessoaRepository pessoaRepository, ParoquiaRepository paroquiaRepository, UsuarioRepository usuarioRepository) {
         this.pessoaParoquiaRepository = pessoaParoquiaRepository;
         this.pessoaParoquiaConverter = pessoaParoquiaConverter;
+        this.pessoaConverter = pessoaConverter;
         this.pessoaRepository = pessoaRepository;
         this.paroquiaRepository = paroquiaRepository;
         this.usuarioRepository = usuarioRepository;
@@ -62,7 +70,10 @@ public class PessoaParoquiaService {
 
     void deleteOldPessoaParoquia(Long paroquiaId) {
         Paroquia paroquia = paroquiaRepository.getOne(paroquiaId);
-        pessoaParoquiaRepository.deleteByPessoaIdAndParoquiaId(paroquia.getUser().getPessoa().getId(), paroquia.getId());
+        Usuario oldUser = paroquia.getUser();
+        if (oldUser != null) {
+            pessoaParoquiaRepository.deleteByPessoaIdAndParoquiaId(paroquia.getUser().getPessoa().getId(), paroquia.getId());
+        }
     }
 
     void deleteOldPessoaParoquiaByPessoaId(Long pessoaId) {
@@ -76,5 +87,11 @@ public class PessoaParoquiaService {
         }
 
         return usuario;
+    }
+
+    public List<PessoaResponse> getAllPessoaByParoquia(Long paroquiaId) {
+        Page<BigInteger> pessoaIds = pessoaParoquiaRepository.findAllPessoaIdsByParoquiaAndName(List.of(paroquiaId), "%%", null);
+        List <Pessoa> pessoas = pessoaRepository.findAllById(pessoaIds.getContent().stream().map(BigInteger::longValue).collect(Collectors.toList()));
+        return pessoaConverter.convert(pessoas);
     }
 }
